backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
python-multipart==0.0.6
openai-whisper==20231117
yt-dlp==2025.3.31
pydub==0.25.1
ffmpeg-python==0.2.0
python-dotenv==1.0.1
openai==0.28.1

# Versioning
semantic-release